WS connecting and login check
Fixed som small buggs and made a WebSocket that connects, and exchanges some information. When the host connects players will be able to join, and the host is remembered for future use. Cleanup required in play.js. API to check if a game is running before connecting to /play. TODO:
make sure that the /play works and switch the WS to there to exchange user information.

Colour selection wheel on the main screen.