Unfinished task assignment
Unfinished commit of all the stuff that I have recently been working on. This is supposed to be able to receive a start signal from the host and then assign tasks to all players. The parts that have been written are unfinished and probably does not work as they are supposed to. As it is currently planned this looks to become inefficient AF so a code optimization is needed as well.

TL;DR: Spaghetti code that is unfinished, inefficient and does not work.