Bug fix 2
There was some more bugs introduced in the update changing up the API-call and readAllFiles. This is now fixed, (hopefully). The async code that checked the tasks before send did not work as intended, unsure why this wasn't caught in testing last time. The QR-generator also used the readAllFiles, and has now been fixed to use the new list format.