Task system first test (#3)
* Redid API and Settingstest

Redid the API and Settignstest on clientside. Now checks if there is more Tasks entered than enabled. Still nothing preventing you from entering the settings wrong in the first place.

* No Tasks overshoot

If a task is disabled it won't count towards the max amount of that task type. When having the max amount of tasks counter reduces to match new max amount.

* convenient setting

Wont reload server when JSON file is saved.

* Bugfix

Fixed but where a invalid taskSettings.json made the /host not load properly. Now it does not rely on the JSON when adaptively making the site content.

* Bug fix 2

There was some more bugs introduced in the update changing up the API-call and readAllFiles. This is now fixed, (hopefully). The async code that checked the tasks before send did not work as intended, unsure why this wasn't caught in testing last time. The QR-generator also used the readAllFiles, and has now been fixed to use the new list format.

* Small fixes

Fixed so that the ws doesn't have a static IP, rather uses the one you are currently connected to. Started on the actions that will take place when the game is started, still not completed.

* Task pt. 1

First idea of how to implement the tasks. HTML does not currently work as there is no doctype. Unsure if this implementation will work, maybe add another funciton that removes all that boilerplate when the file is copied into the publics folder.

* Task pt. 2

Excludes all parameters that just determine HTML settings.