qr rework
Redid the QR system so that only the enabled tasks have to be loaded.

TODO, fixup so that the names of names with multiple spaces all get replaced with "-" when the task name is created.
Fixup so that the files that aren't enabled doesn't get loaded in read_all_files.js