Task EJS support
Now .ejs is supported as well as .html. The ejs Settings can be fetched serverside however one wishes. Uncertain of how useful this will be, but perhaps it could work in someones favor. Once the task system is fully implemented serverside I mean to add a URL to test tasks to see how they render and preform as a part of the reset of the game.

Also included is a improved rework function that redoes the QR-code and replaces all files in the publics folder. If there is any ejs files these are also re-rendered.