less clutter + fix
Removed the checkName property from taskobject as this did not seem that necessary, I could download a list of swears and check towards that but since it is local I do not have to care about who is named what. Also redid checkType to be a IIFE, making it anonymous, making the code more readable and less cluttered. Also fixed the host.js where the "directory" constant that was previously removed was still being used.