Timer + game end
Now when the game starts a timer that counts down towards the end of the game. When this goal is reached the game ends, and that logic is also implemented. Now a lot of animations and visuals have to be done for the game start and game end sequence to make it all look clean.

Counter does now work property as the  property "totalTasks" comes out as undefined, for some reason. That will have to be fixed.