Emergency meeting start
Started with the emergency meeting system, that is toggled, as of right now, only when the button is pressed. It stops the timer and sends to the clients to do the same.

Now I just have to add the Vote UI and make the player get kicked. Another feature is for players to be able to call emergency by reporting a dead body. For this to have happened somebody must have died. After that it is just the ability to do tasks that needs to be implemented and the game will be playable