Clientside markup
Started the ingame actionbar with all the possible actions that can be made in game. All of it should work (have not tested it yet). The QR-part is still not done and needs some work. Currently a issue where the camera function doesn't work unless you are using HTTPS. Uncertain if there is a workaround. The possebility is to use a secure tunnel to a public domain that uses HTTPS, but this could be complicated and/or not work. The other way is to use a self-signed certificate, but using this results in a big fat warning all over the screen when logging into the site, as well as having to specify the https:// part of the URL. Neither is a preferred option but I will have to do something about that issue.