taskcount revition
Redid the taskcount system to include a variable "pushbackTaskcount" that is not updated until a emergency meeting is called either by pressing the button or by reporting a body. This system also adds all of a dead players tasks onto the counter once their body has been reported. Unlike the original game I am considering not having a player do tasks once they're dead not to confuse the live players.

There is still no mechanism that makes you win once all the tasks are done.

This feature adds in functionality to supply a option to not immediately update the taskcount.