Fixed game end logic
All the required systems  that are used when the game ends are now in place, and hopefully work this time. From the tests I have made it works fine. Now the only thing missing is animations. What is left is the startgame, endgame, and votekick animations. After that tasks are required. To work well on phone a CSS rework is due.