Emergency meeting player icons
The player icons during a emergency meeting now has a colour! I also redid the playerside game to instead of having the long line of non-understandable js to have a template literal of HTML. This change is aimed to be able to put in the fetched svg task. It also comes with the benefit of being more readable.

I also did some minor css changes to the "player" to be a grid instead of flex. This makes it centre better. Incorporated these changes to the player on the waiting scree to provide consistency.