Small icons start
Started with the small icons on who voted what when done with a emergency meeting. It should work, fingers crossed, but I haven't had enough time to make sure every edge case works. Also changed hostWS.js to be "defer" instead of having everything in a DOMcontentLoaded listener. This seems to me to be a better way of doing things.

In play.js a few optimizations (unremarkable difference) were made. For example avoiding the multiple spread operators on line 124 and forward. This new system is much more neater to read and should preform better.

In voting things were changed to comply with the new demand for who whated what. Adding to this, a new system were vote amount is indexed is used instead of adding up the sum every single time. Preferable one would use a iterator but I am still contemplating the exact implementations. This probably won't happen as this system works and is easasier to understand.