Tweaked icons
Tweaked the functionality of the mini icons to be animated, so that one pops in every half a second. really cute!

Also fixed some small bugs that was disturbing the normal functionality of the voting system.

Next to do is to make the icons that are put onto the invisible player "skipped vote" show up.